requirements.txt
telethon[cryptg]
pyyaml
aiohttp[speedups]
bs4
